requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.8.0
transformers>=4.35.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Testing and evaluation
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
requests>=2.31.0

# Logging and monitoring
loguru>=0.7.0
wandb>=0.15.0

# Type checking
mypy>=1.5.0
types-requests>=2.31.0